requirements.txt
python-dotenv==0.19.0
aiogram==3.3.0
aiogram-dialog==2.1.0
json-log-formatter==0.5.1 